+++ FenceApp/Screen/LoginView/LoginViewController/ViewTransitionHandler.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+class ViewTransitionHandler {
+    
+    private var window: UIWindow?
+    
+    init(window: UIWindow?) {
+        self.window = window
+    }
+    
+    func transitionToMainView() {
+        let mainViewHandler = MainViewHandler()
+        window?.rootViewController = mainViewHandler.makeTabbarController()
+        window?.makeKeyAndVisible()
+        mainViewHandler.setNavigationControllers()
+    }
+}
